Tile styles and placeholder logic from Crew Cast (ActorPage) are moved to a separate component
