+++ src/features/ActorPage/Content/Crew/index.js
@@ -1,23 +1,10 @@
+import React from "react";
 import { useNavigate } from "react-router-dom";
 import { toMovie } from "../../../../routes";
-import { useSelector } from 'react-redux';
-import { selectCrew } from '../../actorDetailsSlice';
-import {
-    StyledTile,
-    MovieInfo,
-    MovieTitle,
-    MovieSubTitle,
-    RatingWrapper,
-    RatingStar,
-    RatingNumber,
-    VoteCount,
-    MoviePoster,
-    MoviesSection,
-    Title,
-    Rating
-} from "./styled";
-import { GenresList } from "../../../../common/GenresList";
-import { getYear } from "../../../../functions/getYear";
+import { useSelector } from "react-redux";
+import { selectCrew } from "../../actorDetailsSlice";
+import { MoviesSection, Title } from "./styled";
+import { MovieTile } from "../../../../common/MovieTile";
 
 export const Crew = () => {
     const crew = useSelector(selectCrew);
@@ -29,34 +16,14 @@ export const Crew = () => {
 
     return (
         <MoviesSection>
-            <Title>Movies - crew</Title>
-            {crew.map((movie, index) => {
-                const posterUrl = movie.poster_path ? `https://image.tmdb.org/t/p/w500${movie.poster_path}` : null;
-
-                return (
-                    <StyledTile key={index} onClick={() => handleMovieClick(movie.id)}>
-                        {posterUrl ? (
-                            <MoviePoster src={posterUrl} alt={movie.title} />
-                        ) : (
-                            <MoviePoster as="div" $placeholderPoster />
-                        )}
-                        <MovieInfo>
-                            <MovieTitle>{movie.title}</MovieTitle>
-                            <MovieSubTitle>{getYear(movie.release_date)}</MovieSubTitle>
-                            <GenresList
-                                genresIds={movie.genre_ids}
-                            />
-                            <RatingWrapper>
-                                <Rating>
-                                    <RatingStar />
-                                    <RatingNumber>{movie.vote_average.toFixed(1)}</RatingNumber>
-                                </Rating>
-                                <VoteCount>{movie.vote_count} votes</VoteCount>
-                            </RatingWrapper>
-                        </MovieInfo>
-                    </StyledTile>
-                );
-            })}
+            <Title>Crew - Movies</Title>
+            {crew.map((movie, index) => (
+                <MovieTile
+                    key={index}
+                    movie={movie}
+                    handleMovieClick={handleMovieClick}
+                />
+            ))}
         </MoviesSection>
     );
 };
